src/main/java/com/journey/spring/quick/start/service/UserService.java
package com.journey.spring.quick.start.service;

import com.journey.spring.quick.start.dao.LoginLogDao;
import com.journey.spring.quick.start.dao.UserDao;
import com.journey.spring.quick.start.domain.LoginLog;
import com.journey.spring.quick.start.domain.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

/**
 * PROJECT: spring_test
 * Created by xiaxiangnan on 15-12-17.
 */
@Service //标注为服务层的Bean
public class UserService {

    @Autowired
    private UserDao userDao;

    @Autowired
    private LoginLogDao loginLogDao;


    public boolean hasMatchUser(String userName, String password) {
        int matchCount = userDao.getMatchCount(userName, password);
        return matchCount > 0;
    }

    public User findUserByUserName(String userName) {
        return userDao.findUserByUserName(userName);
    }

    public void loginSuccess(User user) {
        user.setCredits(5 + user.getCredits());
        LoginLog loginLog = new LoginLog();
        loginLog.setUserId(user.getUserId());
        loginLog.setIp(user.getLastIp());
        loginLog.setLoginDate(user.getLastVisit());
        userDao.updateLoginInfo(user);
        loginLogDao.insertLoginLog(loginLog);
    }

}
